requirements.txt
edge-tts==6.1.3
lingua-language-detector==1.3.2
Flask==2.2.3